+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
+prisma/schema.prisma
 dist
 coverage
 data
