+++ package.json
@@ -55,12 +55,15 @@
     "typescript": "^5.3.3"
   },
   "dependencies": {
+    "@prisma/client": "^5.13.0",
     "bcrypt": "^5.1.1",
     "cli-color": "^2.0.3",
+    "D": "^1.0.0",
     "express": "^4.18.2",
     "fast-glob": "^3.3.2",
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.3.0",
+    "prisma": "^5.13.0",
     "validator": "^13.11.0"
   },
   "engines": {
