+++ src/infrastructure/repositories/mongodb/LogRepository.ts
@@ -1,7 +1,7 @@
 import { type ILogErrorRepository } from '@/core/ports/driven'
 import { MongoDBHelper } from '.'
 
-export class LogMongoRepository implements ILogErrorRepository {
+export class LogRepository implements ILogErrorRepository {
   async logError (stack: string): Promise<void> {
     const collection = MongoDBHelper.getCollection('errors')
     await collection.insertOne({
