+++ src/infrastructure/repositories/mongodb/index.ts
@@ -1,5 +1,5 @@
 export * from './MongoDBHelper'
-export * from './LogMongoRepository'
-export * from './OrderMongoRepository'
-export * from './AccountMongoRepository'
-export * from './ProductMongoRepository'
+export * from './LogRepository'
+export * from './OrderRepository'
+export * from './AccountRepository'
+export * from './ProductRepository'
