+++ src/infrastructure/repositories/prismaClient.ts
@@ -1,0 +1,3 @@
+import { PrismaClient } from '@prisma/client'
+
+export const prismaClient = new PrismaClient()
