+++ src/main/factories/decorators/makeLogControllerDecorator.ts
@@ -1,8 +1,8 @@
-import { LogMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { LogRepository } from '@/infrastructure/repositories/mongodb'
 import { LogControllerDecorator } from '@/main/decorators'
 import { type IController } from '@/core/ports/driving/presentation'
 
 export const makeLogControllerDecorator = (controller: IController): IController => {
-  const logMongoRepository = new LogMongoRepository()
+  const logMongoRepository = new LogRepository()
   return new LogControllerDecorator(controller, logMongoRepository)
 }
