+++ src/main/factories/usecases/makeAddOrder.ts
@@ -1,8 +1,8 @@
 import { type IAddOrder } from '@/core/ports/driving/services'
 import { AddOrder } from '@/application/services'
-import { OrderMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { OrderRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbAddOrder = (): IAddOrder => {
-  const repository = new OrderMongoRepository()
+  const repository = new OrderRepository()
   return new AddOrder(repository)
 }
