+++ src/main/factories/usecases/makeAddProduct.ts
@@ -1,8 +1,8 @@
 import { type IAddProduct } from '@/core/ports/driving/services'
 import { AddProduct } from '@/application/services'
-import { ProductMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { ProductRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbAddProduct = (): IAddProduct => {
-  const repository = new ProductMongoRepository()
+  const repository = new ProductRepository()
   return new AddProduct(repository)
 }
