+++ src/main/factories/usecases/makeDeleteProduct.ts
@@ -1,8 +1,8 @@
 import { DeleteProduct } from '@/application/services'
 import { type IDeleteProduct } from '@/core/ports/driving/services'
-import { ProductMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { ProductRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbDeleteProduct = (): IDeleteProduct => {
-  const repository = new ProductMongoRepository()
+  const repository = new ProductRepository()
   return new DeleteProduct(repository)
 }
