+++ src/main/factories/usecases/makeLoadOrders.ts
@@ -1,8 +1,8 @@
 import { type ILoadOrders } from '@/core/ports/driving/services'
 import { LoadOrders } from '@/application/services'
-import { OrderMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { OrderRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbLoadOrders = (): ILoadOrders => {
-  const repository = new OrderMongoRepository()
+  const repository = new OrderRepository()
   return new LoadOrders(repository)
 }
