+++ src/main/factories/usecases/makeLoadProductById.ts
@@ -1,8 +1,8 @@
 import { type ILoadProductById } from '@/core/ports/driving/services'
 import { LoadProductById } from '@/application/services'
-import { ProductMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { ProductRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbLoadProductById = (): ILoadProductById => {
-  const repository = new ProductMongoRepository()
+  const repository = new ProductRepository()
   return new LoadProductById(repository)
 }
