+++ src/main/factories/usecases/makeLoadProducts.ts
@@ -1,8 +1,8 @@
 import { type ILoadProducts } from '@/core/ports/driving/services'
 import { LoadProducts } from '@/application/services'
-import { ProductMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { ProductRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbLoadProducts = (): ILoadProducts => {
-  const repository = new ProductMongoRepository()
+  const repository = new ProductRepository()
   return new LoadProducts(repository)
 }
