+++ src/main/factories/usecases/makeLogout.ts
@@ -1,8 +1,8 @@
 import { type ILogout } from '@/core/ports/driving/services'
 import { Logout } from '@/application/services'
-import { AccountMongoRepository } from '@/infrastructure/repositories'
+import { AccountRepository } from '@/infrastructure/repositories'
 
 export const makeDbLogout = (): ILogout => {
-  const repository = new AccountMongoRepository()
+  const repository = new AccountRepository()
   return new Logout(repository)
 }
