+++ src/main/factories/usecases/makeUpdateOrder.ts
@@ -1,8 +1,8 @@
 import { type IUpdateOrder } from '@/core/ports/driving/services'
 import { UpdateOrder } from '@/application/services'
-import { OrderMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { OrderRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbUpdateOrder = (): IUpdateOrder => {
-  const repository = new OrderMongoRepository()
+  const repository = new OrderRepository()
   return new UpdateOrder(repository)
 }
