+++ src/main/factories/usecases/makeUpdateProduct.ts
@@ -1,8 +1,8 @@
 import { type IUpdateProduct } from '@/core/ports/driving/services'
 import { UpdateProduct } from '@/application/services'
-import { ProductMongoRepository } from '@/infrastructure/repositories/mongodb'
+import { ProductRepository } from '@/infrastructure/repositories/mongodb'
 
 export const makeDbUpdateProduct = (): IUpdateProduct => {
-  const repository = new ProductMongoRepository()
+  const repository = new ProductRepository()
   return new UpdateProduct(repository)
 }
