+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "lib": ["ESNext"],
     "strictNullChecks": false,
     "outDir": "dist",
     "module": "commonJS",
